Support submitting SSN in addition to aktørID
When handling requests to search for cases, will perform a validation of
input to see if it is a valid norwegian national identity number.

If the input is a valid NNIN, will exchange it to an aktørID by calling
Aktørregisteret. Otherwise input is treated as an aktørID, as before.