Avoid NNINs in URLs
To be in compliance with current security policy we avoid submitting
NNINs as URL fragments. Logging NNIN in access logs is frowned upon.

I wanted to name the header 'NAV-Person-ID', but the fetch API lower
cases header names, so in order to avoid confusion I chose to refer to
it using lower case everywhere.