Use separate routers for each part of the API.
Due to timing issues the error handling of setting up the S3 connection
was changed, from not defining the feedback routes if S3 is not
available to always defining the routes, and just let them fail with 500
if S3 is not available. It was deemed more natural to let the behaviour
of the API change instead of the structure of the API in failure modes.

The timing issue: when waiting for the S3 promise to resolve before
setting up the routes, the static routes would be defined earlier, thus
making the feedback routes unreachable. I later discovered that one can
get around this issue by adding the router ('app.use(<url>, router)')
before adding the static routes, and subsequently define the actual
routes on the router ('router.get(<url>, <handler>)'), but 500 is a
better error code than 404 for the error situation.