Support reloading the page
Store the ID of the current behandling in session storage, in order to
be able to show the same behandling if the page is reloaded.

In addition to being good web practice to support reload, this helps
when developing, as we no longer need to search for
behandling(er) after Parcel has done some hot reloading of updated code.

This could/should be replaced by putting the ID of the behandling in the
URL.